feat: improve create-react-component script
- support onlyNamedExport option
- bumpup assert package
- remove mock files and it was replaced by snapshotTest
- split files
