+++ gen-ssp/gen/pbjs.ts
@@ -1,5 +1,5 @@
 import { GenOptions } from "./index.ts";
-import * as path from "https://deno.land/std@0.126.0/path/mod.ts";
+import * as path from "https://deno.land/std@0.138.0/path/mod.ts";
 
 const __dirname = path.dirname(path.fromFileUrl(import.meta.url));
