+++ gen-ssp/pbjs/misc.ts
@@ -1,5 +1,5 @@
-import { ensureDir } from "https://deno.land/std@0.126.0/fs/mod.ts";
-import * as path from "https://deno.land/std@0.126.0/path/mod.ts";
+import { ensureDir } from "https://deno.land/std@0.138.0/fs/mod.ts";
+import * as path from "https://deno.land/std@0.138.0/path/mod.ts";
 
 export interface TextFile {
   dir: string[];
