+++ keycloak/oidc-auth-for-cli.ts
@@ -1,4 +1,4 @@
-import { serve } from "https://deno.land/std@0.126.0/http/server.ts";
+import { serve } from "https://deno.land/std@0.138.0/http/server.ts";
 import { Command } from "https://deno.land/x/cliffy@v0.20.1/command/mod.ts";
 import openBrowser from "../misc/open-browser.ts";
