+++ www-snippet/create-react-component/CreateReactComponentOption.ts
@@ -1,0 +1,4 @@
+export interface CreateReactComponentOption {
+  forwardRef?: boolean;
+  onlyNamedExport?: boolean;
+}
