--- www-snippet/create-react-component/__mocks__/react-component/LoremIpsumWithRef_tsx
@@ -1,14 +1,0 @@
-import React, { forwardRef, memo } from "react";
-
-type RefElement = HTMLDivElement;
-type RefElementProps = React.HTMLAttributes<HTMLDivElement>;
-
-export interface LoremIpsumProps extends RefElementProps {}
-
-const LoremIpsum = forwardRef<RefElement, LoremIpsumProps>((props, ref) => {
-  return <div ref={ref} {...props}>Lorem Ipsum</div>;
-});
-
-export default memo(LoremIpsum);
-
-LoremIpsum.displayName = "LoremIpsum";
