--- www-snippet/create-react-component/__mocks__/react-component/LoremIpsum_spec_tsx
@@ -1,11 +1,0 @@
-import { render } from "@testing-library/react";
-
-import LoremIpsum from ".";
-
-describe("<LoremIpsum />", () => {
-  it("should render children correctly", () => {
-    const content = "Lorem Ipsum";
-    const { queryByText } = render(<LoremIpsum>{content}</LoremIpsum>);
-    expect(queryByText("Lorem Ipsum")).toBeInTheDocument();
-  });
-});
