--- www-snippet/create-react-component/__mocks__/react-component/LoremIpsum_tsx
@@ -1,8 +1,0 @@
-import React, { memo, PropsWithChildren } from "react";
-export interface LoremIpsumProps {}
-
-const LoremIpsum = (({}: PropsWithChildren<LoremIpsumProps>) => {
-  return <>LoremIpsum</>;
-});
-
-export default memo(LoremIpsum);
