--- www-snippet/create-react-component/__mocks__/react-component/index_ts
@@ -1,2 +1,0 @@
-export { default } from "./LoremIpsum";
-export * from "./LoremIpsum";
