--- www-snippet/create-react-component/__mocks__/readMockFile.test.ts
@@ -1,10 +1,0 @@
-import { mockFilePaths, readMockFile } from "./readMockFile.ts";
-
-Deno.test(
-  `should exist every mock files`,
-  async () => {
-    for (const filePath of mockFilePaths) {
-      await readMockFile(filePath);
-    }
-  },
-);
